--- programs/scope/src/utils/ctokens/math/mod.rs
@@ -1,9 +1,0 @@
-//! Math for preserving precision
-
-mod common;
-mod decimal;
-mod rate;
-
-pub use common::*;
-pub use decimal::*;
-pub use rate::*;
